+++ docs/source/changelog.rst
@@ -5,6 +5,7 @@ Change Log
 **Pyro 4.28**
 
 - implemented dir() on a Proxy to also return remote methods if known (useful for autocompletion in certain python shells)
+- USE_MSG_WAITALL config item added because there remain certain other systems where MSG_WAITALL is unreliable
 
 
 **Pyro 4.27**